@@ -1,11 +1,10 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.github.dcendents.android-maven'
-
-group='com.github.KalebKE'
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 34
     buildToolsVersion "29.0.3"
+
+    namespace = "com.kircherelectronics.fsensor"
 
     lintOptions {
         abortOnError false
@@ -13,7 +12,7 @@ android {
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 29
+        targetSdkVersion 33
         versionCode 9
         versionName "2.2"
 
@@ -37,5 +36,5 @@ dependencies {
     })
     testImplementation 'junit:junit:4.12'
 
-    api files('libs/commons-math3-3.6.1.jar')
+    implementation 'org.apache.commons:commons-math3:3.6.1'
 }
