@@ -1,20 +1,22 @@
 package com.kircherelectronics.fsensor.sensor.acceleration;
 
-import android.content.Context;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 
-import com.kircherelectronics.fsensor.filter.gyroscope.fusion.complementary.OrientationFusedComplementary;
-import com.kircherelectronics.fsensor.linearacceleration.LinearAcceleration;
-import com.kircherelectronics.fsensor.linearacceleration.LinearAccelerationFusion;
-import com.kircherelectronics.fsensor.observer.SensorSubject;
+import com.kircherelectronics.fsensor.sensor.orientation.fusion.fusion.complementary.ComplimentaryOrientation;
 import com.kircherelectronics.fsensor.sensor.FSensor;
+import com.kircherelectronics.fsensor.sensor.FSensorEvent;
+import com.kircherelectronics.fsensor.sensor.FSensorEventListener;
+import com.kircherelectronics.fsensor.util.gravity.GravityUtil;
 import com.kircherelectronics.fsensor.util.rotation.RotationUtil;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /*
- * Copyright 2018, Kircher Electronics, LLC
+ * Copyright 2024, Tracqi Technology, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,208 +35,101 @@ public class ComplementaryLinearAccelerationSensor implements FSensor {
     private static final String TAG = ComplementaryLinearAccelerationSensor.class.getSimpleName();
 
     private final SensorManager sensorManager;
-    private final SimpleSensorListener listener;
-    private float startTime = 0;
-    private int count = 0;
+    private final List<FSensorEventListener> fSensorEventListeners = new ArrayList<>();
+    private final SensorEventListener sensorEventListener = new SensorListener();
 
-    private float[] magnetic = new float[3];
-    private float[] rawAcceleration = new float[3];
-    private float[] rotation = new float[3];
-    private float[] acceleration = new float[3];
-    private float[] output = new float[4];
+    private final float[] acceleration = new float[3];
+    private final float[] magnetic = new float[3];
+    private final float[] rotation = new float[3];
+    private final float[] output = new float[4];
+    private long rotationTimestamp;
 
-    private LinearAcceleration linearAccelerationFilterComplimentary;
-    private OrientationFusedComplementary orientationFusionComplimentary;
+    private final ComplimentaryOrientation orientationFusionComplimentary = new ComplimentaryOrientation();
 
-    private int sensorDelay = SensorManager.SENSOR_DELAY_FASTEST;
-    private int sensorType = Sensor.TYPE_GYROSCOPE;
-
-    private final SensorSubject sensorSubject;
-
-    public ComplementaryLinearAccelerationSensor(Context context) {
-        this.sensorManager = (SensorManager) context.getSystemService(Context.SENSOR_SERVICE);
-        this.listener = new SimpleSensorListener();
-        this.sensorSubject = new SensorSubject();
-        initializeFSensorFusions();
-    }
-
-    /**
-     * Start the sensor.
-     */
-    @Override
-    public void start() {
-        startTime = 0;
-        count = 0;
-        registerSensors(sensorDelay);
-    }
-
-    /**
-     * Stop the sensor.
-     */
-    @Override
-    public void stop() {
-        unregisterSensors();
+    public ComplementaryLinearAccelerationSensor(SensorManager sensorManager) {
+        this.sensorManager = sensorManager;
     }
 
     @Override
-    public void register(SensorSubject.SensorObserver sensorObserver) {
-        sensorSubject.register(sensorObserver);
-    }
-
-    @Override
-    public void unregister(SensorSubject.SensorObserver sensorObserver) {
-        sensorSubject.unregister(sensorObserver);
-    }
-
-    /**
-     * Set the gyroscope sensor type.
-     * @param sensorType must be Sensor.TYPE_GYROSCOPE or Sensor.TYPE_GYROSCOPE_UNCALIBRATED
-     */
-    public void setSensorType(int sensorType) {
-        if(sensorType != Sensor.TYPE_GYROSCOPE && sensorType != Sensor.TYPE_GYROSCOPE_UNCALIBRATED) {
-            throw new IllegalStateException("Sensor Type must be Sensor.TYPE_GYROSCOPE or Sensor.TYPE_GYROSCOPE_UNCALIBRATED");
+    public void registerListener(FSensorEventListener sensorEventListener, int sensorDelay) {
+        if (fSensorEventListeners.isEmpty()) {
+            registerSensors(sensorDelay);
         }
 
-        this.sensorType = sensorType;
-
+        fSensorEventListeners.add(sensorEventListener);
     }
 
-    /**
-     * Set the sensor frequency.
-     * @param sensorDelay Must be SensorManager.SENSOR_DELAY_FASTEST, SensorManager.SENSOR_DELAY_GAME, SensorManager.SENSOR_DELAY_NORMAL or SensorManager.SENSOR_DELAY_UI
-     */
-    public void setSensorDelay(int sensorDelay) {
-        if(sensorDelay != SensorManager.SENSOR_DELAY_FASTEST && sensorDelay != SensorManager.SENSOR_DELAY_GAME && sensorDelay != SensorManager.SENSOR_DELAY_NORMAL && sensorDelay != SensorManager.SENSOR_DELAY_UI) {
-            throw new IllegalStateException("Sensor Frequency must be SensorManager.SENSOR_DELAY_FASTEST, SensorManager.SENSOR_DELAY_GAME, SensorManager.SENSOR_DELAY_NORMAL or " +
-                    "SensorManager.SENSOR_DELAY_UI");
+    @Override
+    public void unregisterListener(FSensorEventListener sensorEventListener) {
+        this.fSensorEventListeners.remove(sensorEventListener);
+        if (fSensorEventListeners.isEmpty()) {
+            unregisterSensors();
         }
-        this.sensorDelay = sensorDelay;
     }
 
-    /**
-     * The filter coefficient.
-     *
-     * @param timeConstant A floating point value between 0 and 1. Exclusive of 0, inclusive of 1.
-     */
-    public void setFSensorComplimentaryLinearAccelerationTimeConstant(float timeConstant) {
-        orientationFusionComplimentary.setTimeConstant(timeConstant);
+    private void copyAcceleration(float[] rawAcceleration) {
+        System.arraycopy(rawAcceleration, 0, this.acceleration, 0, this.acceleration.length);
     }
 
-    /**
-     * Reset the sensor.
-     */
-    public void reset() {
-        stop();
-        magnetic = new float[3];
-        acceleration = new float[3];
-        rawAcceleration = new float[3];
-        rotation = new float[3];
-        output = new float[4];
-        listener.reset();
-        start();
-    }
-
-    private float calculateSensorFrequency() {
-        // Initialize the start time.
-        if (startTime == 0) {
-            startTime = System.nanoTime();
-        }
-
-        long timestamp = System.nanoTime();
-
-        // Find the sample period (between updates) and convert from
-        // nanoseconds to seconds. Note that the sensor delivery rates can
-        // individually vary by a relatively large time frame, so we use an
-        // averaging technique with the number of sensor updates to
-        // determine the delivery rate.
-
-        return (count++ / ((timestamp - startTime) / 1000000000.0f));
-    }
-
-    private void initializeFSensorFusions() {
-        orientationFusionComplimentary = new OrientationFusedComplementary();
-        linearAccelerationFilterComplimentary = new LinearAccelerationFusion(orientationFusionComplimentary);
-    }
-
-    private void processRawAcceleration(float[] rawAcceleration) {
-        System.arraycopy(rawAcceleration, 0, this.rawAcceleration, 0, this.rawAcceleration.length);
-    }
-
-    private void processAcceleration(float[] acceleration) {
-        System.arraycopy(acceleration, 0, this.acceleration, 0, this.acceleration.length);
-    }
-
-    private void processMagnetic(float[] magnetic) {
+    private void copyMagnetic(float[] magnetic) {
         System.arraycopy(magnetic, 0, this.magnetic, 0, this.magnetic.length);
     }
 
-    private void processRotation(float[] rotation) {
+    private void copyRotation(float[] rotation) {
         System.arraycopy(rotation, 0, this.rotation, 0, this.rotation.length);
     }
 
     private void registerSensors(int sensorDelay) {
-
         orientationFusionComplimentary.reset();
-
-
-        // Register for sensor updates.
-        sensorManager.registerListener(listener, sensorManager
-                        .getDefaultSensor(Sensor.TYPE_ACCELEROMETER),
-                sensorDelay);
-
-
-        // Register for sensor updates.
-        sensorManager.registerListener(listener, sensorManager
-                        .getDefaultSensor(Sensor.TYPE_MAGNETIC_FIELD),
-                sensorDelay);
-
-        // Register for sensor updates.
-        sensorManager.registerListener(listener,
-                sensorManager.getDefaultSensor(sensorType),
-                sensorDelay);
-
+        sensorManager.registerListener(sensorEventListener, sensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER), sensorDelay);
+        sensorManager.registerListener(sensorEventListener, sensorManager.getDefaultSensor(Sensor.TYPE_MAGNETIC_FIELD), sensorDelay);
+        sensorManager.registerListener(sensorEventListener, sensorManager.getDefaultSensor(Sensor.TYPE_GYROSCOPE), sensorDelay);
     }
 
     private void unregisterSensors() {
-        sensorManager.unregisterListener(listener);
+        sensorManager.unregisterListener(sensorEventListener);
     }
 
-    private void setOutput(float[] value) {
-        System.arraycopy(value, 0, output, 0, value.length);
-        output[3] = calculateSensorFrequency();
-        sensorSubject.onNext(output);
+    private void calculateLinerAcceleration(float[] acceleration, float[] gravity) {
+        // Determine the linear acceleration
+        output[0] = acceleration[0] - gravity[0];
+        output[1] = acceleration[1] - gravity[1];
+        output[2] = acceleration[2] - gravity[2];
     }
 
-    private class SimpleSensorListener implements SensorEventListener {
-
+    private class SensorListener implements SensorEventListener {
+        private boolean hasAcceleration = false;
         private boolean hasRotation = false;
         private boolean hasMagnetic = false;
-
-        private void reset() {
-            hasRotation = false;
-            hasMagnetic = false;
-        }
 
         @Override
         public void onSensorChanged(SensorEvent event) {
             if (event.sensor.getType() == Sensor.TYPE_ACCELEROMETER) {
-                processRawAcceleration(event.values);
-                if (!orientationFusionComplimentary.isBaseOrientationSet()) {
-                    if (hasRotation && hasMagnetic) {
-                        orientationFusionComplimentary.setBaseOrientation(RotationUtil.getOrientationVectorFromAccelerationMagnetic(rawAcceleration, magnetic));
-                    }
-                } else {
-                    orientationFusionComplimentary.calculateFusedOrientation(rotation, event.timestamp, rawAcceleration, magnetic);
-                    processAcceleration(linearAccelerationFilterComplimentary.filter(rawAcceleration));
-
-                    setOutput(acceleration);
-                }
+                copyAcceleration(event.values);
+                hasAcceleration = true;
             } else if (event.sensor.getType() == Sensor.TYPE_MAGNETIC_FIELD) {
-                processMagnetic(event.values);
+                copyMagnetic(event.values);
                 hasMagnetic = true;
-            } else if (event.sensor.getType() == sensorType) {
-                processRotation(event.values);
+            } else if (event.sensor.getType() == Sensor.TYPE_GYROSCOPE) {
+                copyRotation(event.values);
+                rotationTimestamp = event.timestamp;
                 hasRotation = true;
+            }
+
+            if (hasAcceleration && hasRotation && hasMagnetic) {
+                hasAcceleration = false;
+                hasRotation = false;
+                hasMagnetic = false;
+
+                if (!orientationFusionComplimentary.isBaseOrientationSet()) {
+                    orientationFusionComplimentary.setBaseOrientation(RotationUtil.getOrientationVector(acceleration, magnetic));
+                } else {
+                    float[] gravity = GravityUtil.getGravityFromOrientation(orientationFusionComplimentary.calculateFusedOrientation(rotation, rotationTimestamp, acceleration, magnetic));
+                    calculateLinerAcceleration(event.values, gravity);
+                    for (FSensorEventListener fSensorEventListeners : fSensorEventListeners) {
+                        fSensorEventListeners.onSensorChanged(new FSensorEvent(event.sensor, event.accuracy, event.timestamp, output));
+                    }
+                }
             }
         }
 
