@@ -1,4 +1,4 @@
-package com.kircherelectronics.fsensor.filter.gyroscope.fusion.kalman.filter;
+package com.kircherelectronics.fsensor.sensor.orientation.fusion.fusion.kalman.filter;
 
 import org.apache.commons.math3.filter.ProcessModel;
 import org.apache.commons.math3.linear.Array2DRowRealMatrix;
